Move `BaseCommand` from actions.ts to base.ts
That's surely where it belongs, and in general, I'd like to split actions.ts up.
